Run pytest and sphinx in parallel
