+++ hdl_registers/__init__.py
@@ -18,6 +18,7 @@ REPO_ROOT = Path(__file__).parent.parent.resolve()
 HDL_REGISTERS_PATH = REPO_ROOT / "hdl_registers"
 HDL_REGISTERS_DOC = REPO_ROOT / "doc"
 HDL_REGISTERS_GENERATED = REPO_ROOT / "generated"
+HDL_REGISTERS_GENERATED_SPHINX_HTML = HDL_REGISTERS_GENERATED / "sphinx_html"
 HDL_REGISTERS_TESTS = REPO_ROOT / "tests"
 HDL_REGISTERS_TOOLS = REPO_ROOT / "tools"
